refractor: create component flyout
